refactor(architecture): Step 4 - Dissolve utils directory and re-home logic to state and telemetry
